Renamed grafana controller objects
