+++ pkg/client/clientset/versioned/typed/grafana/v1alpha1/fake/fake_grafanadashboard.go
@@ -19,7 +19,7 @@ limitations under the License.
 package fake
 
 import (
-	v1alpha1 "kubernetes-grafana-controller/pkg/apis/samplecontroller/v1alpha1"
+	v1alpha1 "kubernetes-grafana-controller/pkg/apis/grafana/v1alpha1"
 
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
