+++ pkg/client/clientset/versioned/typed/grafana/v1alpha1/grafanadashboard.go
@@ -19,7 +19,7 @@ limitations under the License.
 package v1alpha1
 
 import (
-	v1alpha1 "kubernetes-grafana-controller/pkg/apis/samplecontroller/v1alpha1"
+	v1alpha1 "kubernetes-grafana-controller/pkg/apis/grafana/v1alpha1"
 	scheme "kubernetes-grafana-controller/pkg/client/clientset/versioned/scheme"
 	"time"
