+++ pkg/client/informers/externalversions/generic.go
@@ -20,7 +20,7 @@ package externalversions
 
 import (
 	"fmt"
-	v1alpha1 "kubernetes-grafana-controller/pkg/apis/samplecontroller/v1alpha1"
+	v1alpha1 "kubernetes-grafana-controller/pkg/apis/grafana/v1alpha1"
 
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
 	cache "k8s.io/client-go/tools/cache"
