+++ pkg/client/informers/externalversions/grafana/interface.go
@@ -19,8 +19,8 @@ limitations under the License.
 package samplecontroller
 
 import (
+	v1alpha1 "kubernetes-grafana-controller/pkg/client/informers/externalversions/grafana/v1alpha1"
 	internalinterfaces "kubernetes-grafana-controller/pkg/client/informers/externalversions/internalinterfaces"
-	v1alpha1 "kubernetes-grafana-controller/pkg/client/informers/externalversions/samplecontroller/v1alpha1"
 )
 
 // Interface provides access to each of this group's versions.
